disconnected bones, more bone layers, some scaling
Disabled the 'connected' checkbox so that the bones can be moved around, and moved the face and fingers+toes to separate layers. Also renamed the layers via the plugin.

Added a copy scale constraint to the hands, toes, and jaw so that they can be scaled somewhat, tho this isn't the most professional implementation of it since it also scales children. Will look into an easier scale control mechanism that is unreal-compatible in time.