Fixed the rotations
The reason for misbehaving rotations is that the hand controllers were mistakenly set to YZX instead of Quaternion. I painstakingly converted every keyframe on these controllers to Quaternion and changed the controllers on the rig file to quaternion (as well as the forearms, just in case), so there should be no more issues from that from now on.

I also deleted the Y-Location in-between keyframes for the linearly interpolated part of the idle patrol walk to get it in perfect sync with the ground in Unreal.