Rideable cloudrunner files and sound files
Migrated the rideable cloudrunner files into the repo. They're not well-organized with their animations but its fine since they're only used once, and they're there in case we want to make changes to them. We probably won't need them though as cloudrunner riding doesn't happen anywhere else in the entire game IIRC.

Watch out, it seems like I stupidly had the sound files for the win cutscene in my downloads folder, which is now cleared. Grabbed the main track from Unreal.

Also put the sound files for the intro cutscene in. Two appeared to be missing. One was a krystal voice line and the other was a fire sound, but I can't seem to find where these two sounds were used. Probably also from the downloads folder. Its probably fine anyway since it's unlikely we'll need to make changes to that specific cutscene that would require adjusting the timing of the audio.